存在しないidでの削除についての実装。

+++ src/main/java/com/examination/application/DeleteEmployeeUseCase.java
@@ -1,5 +1,6 @@
 package com.examination.application;
 
+import com.examination.application.exception.EmployeeNotFoundException;
 import com.examination.domain.EmployeeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -10,6 +11,9 @@ public class DeleteEmployeeUseCase {
   private final EmployeeRepository employeeRepository;
 
   public void deleteEmployee(String id) {
+    if (employeeRepository.getEmployeeById(id).isEmpty()) {
+      throw new EmployeeNotFoundException(id);
+    }
     employeeRepository.deleteEmployee(id);
   }
 }
